Added Bram's prefered implication arrows
"Also, for the material conditional please use -> instead of => … the => is typically used as a meta-logical symbol of logical implication. The -> is the standard logic symbol for material implication.  I know, plenty of books use => for material implication, but I find that confusing: any use of double lines to me denotes a logical/semantical consequence" -Bram in an Email